Set the .zsh-eza function to return 0 instead of 1 when terminal is dumb
When a TTY is not available, or otherwise $TERM is set to `dumb`, returning 1 can cause the calling application to assume actual failure, potentially interfering with regular function. Returning 0 may be a bit less clear about zsh-eza's failure, but it doesn't misrepresent the state of the shell otherwise.

Printing a message will still communicate that the shell did not load zsh-eza in those cases where the user is trying to figure out where the issue lies.

Signed-off-by: Mikrz <github@mikrz.ca>